requirements.txt
# ==========================================
# Requirements.txt pour Railway
# ViT+MLP Model avec Firebase Integration
# ==========================================

# Flask & Server
Flask==3.0.3
flask-cors==4.0.0
gunicorn==23.0.0

# Computer Vision & ML
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
Pillow==10.3.0
numpy==1.26.4

# PyTorch & Vision Models
torch==2.4.0
torchvision==0.19.0
timm==1.0.3

# Image Augmentation
albumentations==1.4.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.4.2

# Firebase / Google Cloud
google-cloud-storage==2.18.2
google-cloud-firestore==2.16.1
firebase-admin==6.5.0

# Utilities
python-dateutil==2.9.0